#1054: extract poll_binding_process_descriptor to afxdp/poll_descriptor.rs
afxdp.rs went from 3,843 prod LOC to 1,454 by relocating the
2,382-line `poll_binding_process_descriptor` fn (62% of the file)
into a dedicated child module `afxdp/poll_descriptor.rs`.

Visibility widened from file-private to `pub(super)` so afxdp.rs
can still call it; everything else stays unchanged. The new module
gets `use super::*;` to inherit afxdp's full namespace including
sibling submodules (`parser`, `rst`, `sharded_neighbor`,
`shared_ops`, `bind`, `mpsc_inbox`) that the extracted fn references
directly.

Pure relocation — no production logic touched. The 2,382-line fn
body remains an unwieldy state-machine and is a deeper refactor
target tracked separately (overlaps with #963's Builder/Editor
proposal). This PR is purely about file modularity.

LOC distribution after:
- afxdp.rs:                   1,454 (was 3,843) — under threshold
- afxdp/poll_descriptor.rs:   2,400 (new — fn + doc + use super::*)

cargo test --release: 865 passed, 0 failed.
